requirements.txt
cryptography==2.7
dataclasses==0.6
ecological==2.0.0
Flask==1.1.1
gunicorn==19.9.0
Jinja2==2.10.1
mypy==0.730
mypy-extensions==0.4.2
python-json-logger==0.1.11
PyYAML==5.1.2
python-rapidjson==0.8.0
redis==3.3.8
requests==2.22.0
structlog==19.1.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.24.2
Werkzeug==0.16.0
